backend/requirements.txt
cd backend
cat > requirements.txt << 'EOF'
Flask==2.3.3
Flask-CORS==4.0.0
python-telegram-bot==20.3
python-dotenv==1.0.0
gunicorn==21.2.0
EOF